requirements.txt
astropy==7.0.1
astropy-iers-data==0.2025.2.17.0.34.13
contourpy==1.3.1
corner==2.2.3
cycler==0.12.1
emcee==3.1.6
fonttools==4.56.0
jplephem==2.22
kiwisolver==1.4.8
loguru==0.7.3
matplotlib==3.10.0
nestle==0.2.0
numdifftools==0.9.41
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pint-pulsar==1.1.1
pyerfa==2.0.1.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
tk==0.1.0
tqdm==4.67.1
uncertainties==3.2.2
